Docker-ml-assignement/app_api/requirements_api.txt
# app_api/requirements_api.txt
torch
torchvision
fastapi
uvicorn[standard]
pillow
python-multipart